Fix vLLM instance scheduling when tp size is 1
https://github.com/OpenRLHF/OpenRLHF/pull/677
introduced a change that forces to use ray as the vLLM
distributed_executor_backend even if tp size is 1. While this
works for vLLM versions below 0.7.0, it's no longer the case
starting from 0.7.0, as it will force you to use uni even if
you set explicitly to use ray (it takes me so much time to
debug this and finds out this weird change XD, as there is no
warning information regarding this at all!):

https://github.com/vllm-project/vllm/blob/ad34c0df0f1b26b303a590133685b29e3daad20e/vllm/engine/llm_engine.py#L444-L462

So, we still need to treat this the same way as what we do before
for GPU executor, otherwise, if we have multiple vLLM instances scheduled
on one machine, it will all use GPU 0, which will cause errors.

In addition, now we specify explicitly to use uni when tp size is 1,
in case vLLM changed their mind and moved from this weird behavior later.

Signed-off-by: Hollow Man <hollowman@opensuse.org>