Fix vLLM tp=1 when RAY_EXPERIMENTAL_NOSET_*_VISIBLE_DEVICES is set
If `RAY_EXPERIMENTAL_NOSET_*_VISIBLE_DEVICES` is set, and vLLM is not
using the ray distributed executor backend, we need to take care
of the `CUDA_VISIBLE_DEVICES` by ourselves. Otherwise, all the vLLM
engines will be scheduled to GPU 0.

Note that, similar to
https://github.com/OpenRLHF/OpenRLHF/pull/536#issuecomment-2499865726,
we shouldn't call `ray_noset_visible_devices()` in `LLMRayActor`, otherwise
the indirectly imported deepspeed will break it and make the newly
`CUDA_VISIBLE_DEVICES` not work.

Signed-off-by: Hollow Man <hollowman@opensuse.org>