Support full determinism option (#868)
Fix #745 and #834

The non-determinism mainly comes from the flash attention, as
in the backward pass, by default, flash attention uses atomic
adds for performance. Deterministic backward is implemented
in v2.4.1 but not enabled by default, as it's slightly slower
and uses more memory:
https://github.com/Dao-AILab/flash-attention/commit/732654583c2e640adc012ecb60e460bf19dcd9e3

In addition, we switch the set_seed to transformers' helper
function, to remove code duplication.

Signed-off-by: Hollow Man <hollowman@opensuse.org>