Fix full determinism mode when using vLLM V1 (#932)
Collective RPC is supported now in vLLM (nightly), so we
no longer need to always set `VLLM_ENABLE_V1_MULTIPROCESSING=0`
when we use vLLM V1. https://github.com/vllm-project/vllm/pull/15444

But we still need to set that when we enable the full_determinism
mode, as vLLM does not guarantee the reproducibility of the results
by default, for the sake of performance. We need to turn off
multiprocessing to make the scheduling deterministic. This is
not needed and will be ignored for V0.

Signed-off-by: Hollow Man <hollowman@opensuse.org>