Remove implicit import of deepspeed in vLLM to avoid early CUDA calls (#1019)
Now the new version for the synchronous training mode is broken, as
now `LLMRayActor` inherits `BaseLLMRayActor`, which make it import
all the dependencies in the file as well all the up level `__init__.py`,
and introduces early CUDA calls that's caused by the import of deepspeed.

Similar to:
- https://github.com/OpenRLHF/OpenRLHF/pull/524#issuecomment-2501505023
- https://github.com/OpenRLHF/OpenRLHF/pull/537

So this PR clean up such responsible `__init__.py` files (similar to #537)
to give a ultimate fix.

There are also some alternatives to this PR, such as forcing the users
to set `RAY_EXPERIMENTAL_NOSET_CUDA_VISIBLE_DEVICES`. Another
alternative is to set `PYTORCH_NVML_BASED_CUDA_CHECK=1`, but I don't
think this is a good practice as explained in
https://github.com/deepspeedai/DeepSpeed/pull/6810#issuecomment-2528676331

Signed-off-by: Hollow Man <hollowman@opensuse.org>